Test: Reorganize tests for MarkupParser
To keep things organized we split the big `MarkupParserTest` into
multiple test classes and define a tag `MarkupParser` so that JUnit can
execute only tests related to the parser, if desired.

Additionally we are adding a bunch of helper methods to help with
readability when putting together assertions.
