+++ src/test/java/de/zuellich/offlinewiktionary/core/markup/TokenAssertions.java
@@ -2,6 +2,9 @@ package de.zuellich.offlinewiktionary.core.markup;
 
 import static org.junit.jupiter.api.Assertions.fail;
 
+import java.util.List;
+import java.util.function.Consumer;
+
 public class TokenAssertions {
 
   public static void assertMatchingType(MarkupTokenType expected, MarkupToken token) {
@@ -77,6 +80,58 @@ public class TokenAssertions {
       fail(
           String.format(
               "Expected indent level '%d' but got '%d'", expectedLevel, indentToken.level()));
+    }
+  }
+
+  public static void assertTokensStrict(
+      final List<MarkupToken> tokens, final List<Consumer<MarkupToken>> matchers) {
+    if (tokens.size() != matchers.size()) {
+      fail(
+          String.format(
+              "Token count (%d) doesn't match matchers count (%d).",
+              tokens.size(), matchers.size()));
+    }
+
+    assertTokens(tokens, matchers);
+  }
+
+  public static Consumer<MarkupToken> text(String text) {
+    return (MarkupToken token) -> {
+      assertText(token, text);
+    };
+  }
+
+  public static Consumer<MarkupToken> link(String expectedTextAndLabel) {
+    return (MarkupToken token) -> {
+      assertLink(token, expectedTextAndLabel);
+    };
+  }
+
+  public static Consumer<MarkupToken> indent(int level) {
+    return (MarkupToken token) -> {
+      assertIndent(token, level);
+    };
+  }
+
+  /**
+   * Iterate over the provided tokens and see if the supplied matchers match the corresponding
+   * token. You can supply more tokens than matchers, but not the other way around. If you want to
+   * be precise use {@link TokenAssertions#assertTokensStrict(List, List)}
+   */
+  public static void assertTokens(
+      final List<MarkupToken> tokens, final List<Consumer<MarkupToken>> matchers) {
+    if (matchers.size() > tokens.size()) {
+      fail(
+          String.format(
+              "More matchers than tokens. Got %d tokens, but %d matchers.",
+              tokens.size(), matchers.size()));
+    }
+
+    final int min = Math.min(matchers.size(), tokens.size());
+    for (int i = 0; i < min; i++) {
+      final MarkupToken token = tokens.get(i);
+      final Consumer<MarkupToken> matcher = matchers.get(i);
+      matcher.accept(token);
     }
   }
 }
